pyproject.toml
# Sentry Tool - CLI for querying and managing Sentry issues
# Reference: CODE-PY-001 (Python Project Setup pattern)

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "requests-mock>=1.11.0",
  "ruff>=0.8",
  "mypy>=1.13",
  "pre-commit>=4.0",
  "types-requests>=2.31.0"
]

[project]
name = "sentry-tool"
version = "0.1.0"
description = "CLI tool for querying and managing Sentry issues"
readme = "README.md"
license = "MIT"
authors = [
  {name = "Vlad Korolev", email = "vlad@arsenale.bio"}
]
requires-python = ">=3.13"
dependencies = [
  "requests>=2.31.0",
  "rich>=13.0.0",
  "typer>=0.12.0",
  "sentry-sdk>=2.0.0",
  "structlog>=24.0",
  "tenacity>=8.0.0",
  "pydantic-settings>=2.0.0"
]

[project.urls]
Homepage = "https://github.com/vladistan/sentry-tool"
Repository = "https://github.com/vladistan/sentry-tool"
Issues = "https://github.com/vladistan/sentry-tool/issues"
Source = "https://github.com/vladistan/sentry-tool"
Documentation = "https://github.com/vladistan/sentry-tool#readme"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "requests-mock>=1.11.0",
  "ruff>=0.8",
  "mypy>=1.13",
  "pre-commit>=4.0",
  "types-requests>=2.31.0"
]

[project.scripts]
sentry-tool = "sentry_tool.cli:cli"

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "if settings.DEBUG",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == .__main__.:",
  "class .*\\bProtocol\\):",
  "@(abc\\.)?abstractmethod"
]

[tool.coverage.run]
source = ["src/sentry_tool"]
omit = [
  "*/tests/*",
  "*/test_*.py",
  "*/__about__.py",
  "*/__pycache__/*"
]

[tool.hatch.build.targets.wheel]
packages = ["src/sentry_tool"]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true
exclude = [".venv", "build", "dist"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
  "-v",
  "--cov=sentry_tool",
  "--cov-report=term-missing",
  "--cov-report=html"
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "N", # pep8-naming
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
  "SIM", # flake8-simplify
  "TCH", # flake8-type-checking
  "RUF", # ruff-specific rules
  "ERA", # eradicate
  "PL", # pylint
  "PERF" # performance
]
ignore = ["E501"] # line too long, handled by formatter
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["sentry_tool"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # Allow unused imports in __init__.py
"tests/**/*.py" = ["ARG", "S101"] # Allow unused args and asserts in tests
